Update fix lagging around of pointcloud
Before the timestamp was just taken at the point where the icp started. This led to the pointcloud lagging around in the system.
This fixes this issue by taking the timestamp of one of the actual measurements. Currently the oldest stamp is taken, this might have to be reconsidered in the future.